refactor: Store - invitation 추가 및 파일명 변경
